+++ tests/ImageSharp.Drawing.Benchmarks/Drawing/Rounding.cs
@@ -52,7 +52,7 @@ public class Rounding
             if (verticesLength > 0)
             {
                 ri = vertices.Length - (remainder / 2);
-                float maxIterations = verticesLength / (Vector256<float>.Count * 2);
+                nint maxIterations = verticesLength / (Vector256<float>.Count * 2);
                 ref Vector256<float> sourceBase = ref Unsafe.As<PointF, Vector256<float>>(ref MemoryMarshal.GetReference(vertices));
                 ref Vector256<float> destinationBase = ref Unsafe.As<float, Vector256<float>>(ref MemoryMarshal.GetReference(destination));
 
@@ -99,7 +99,7 @@ public class Rounding
             if (verticesLength > 0)
             {
                 ri = vertices.Length - (remainder / 2);
-                float maxIterations = verticesLength / (Vector128<float>.Count * 2);
+                nint maxIterations = verticesLength / (Vector128<float>.Count * 2);
                 ref Vector128<float> sourceBase = ref Unsafe.As<PointF, Vector128<float>>(ref MemoryMarshal.GetReference(vertices));
                 ref Vector128<float> destinationBase = ref Unsafe.As<float, Vector128<float>>(ref MemoryMarshal.GetReference(destination));
